backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
python-pptx==1.0.2
Werkzeug==3.1.3
requests==2.31.0
urllib3==1.26.18
Pillow==10.3.0
pywin32==306; platform_system == "Windows"
bcrypt
PyJWT
python-dotenv
pymongo==4.6.1
dnspython==2.6.1
gunicorn==22.0.0